Moved the person model into types
